Move network rules update test to its own package
